Update to SABnzbd 1.2.0
The new version requires the jessie-backports version of
python-cryptography since that is the only version that has the
cryptography.x509 package needed for loading SSL certificates.

Use the newer version of OpenSSL from jessie-backports as well to be
more paranoid.

Also switch to the release version of the SABnzbd tarball since it has
the version correctly set in sabnzbd/version.py instead of saying
"unknown [develop]" in the Settings screen.