Allow updating host_whitelist
Setting the HOST_WHITELIST_ENTRIES environment variable will update
SABnzbd's `host_whitelist` setting with the container's hostname and the
string of comma-separated hostnames and FQDNs specified in the
HOST_WHITELIST_ENTRIES variable.

This satisfies SABnzbd's hostname verification and allows accessing the
web interface by any of the entries specified in the whitelist.

Fixes #18